+++ src/ngraph/CMakeLists.txt
@@ -156,8 +156,6 @@ set (SRC
     op/experimental/dyn_replace_slice.hpp
     op/experimental/dyn_reshape.cpp
     op/experimental/dyn_reshape.hpp
-    op/experimental/reshape.cpp
-    op/experimental/reshape.hpp
     op/experimental/dyn_slice.cpp
     op/experimental/dyn_slice.hpp
     op/experimental/generate_mask.cpp
