+++ src/ngraph/ngraph.hpp
@@ -120,7 +120,6 @@ namespace ngraph
 #include "ngraph/op/experimental/generate_mask.hpp"
 #include "ngraph/op/experimental/random_uniform.hpp"
 #include "ngraph/op/experimental/range.hpp"
-#include "ngraph/op/experimental/reshape.hpp"
 #include "ngraph/op/experimental/shape_of.hpp"
 #include "ngraph/op/experimental/tile.hpp"
 #include "ngraph/op/experimental/transpose.hpp"
