+++ src/ngraph/op/reshape.hpp
@@ -105,5 +105,52 @@ namespace ngraph
             Shape m_output_shape;
             bool m_is_transpose{false};
         };
+
+        namespace v1
+        {
+            /// \brief Tensor dynamic reshape operation.
+            ///
+            /// "Converts" an input tensor into a new shape with the same number of elements.
+            /// This op does not touch the actual data. If needed, use Transpose for that purpose.
+            ///
+            class Reshape : public Op
+            {
+            public:
+                NGRAPH_API
+                static constexpr NodeTypeInfo type_info{"DynReshape", 1};
+                const NodeTypeInfo& get_type_info() const override { return type_info; }
+                Reshape() = default;
+                /// \brief Constructs a dynamic reshape operation. This operation does not perform
+                ///        transpose.
+                ///
+                /// \param arg The tensor to be reshaped.
+                /// \param pattern The node that defines output shape pattern.
+                ///        If the input shape is \f$(a_0,\dots,a_{k-1})\f$ then the output shape
+                ///        must
+                ///        be of the form \f$(b_0,\dots,b_{j-1})\f$ where \f$\Pi(a_i) = \Pi(b_i)\f$.
+                ///        A value of -1 is allowed for at most one dimension, in which case the
+                ///        dimension size is inferred based on element count of input tensor.
+                /// \param zero_flag Treats zeros in `pattern` as wildcard flags indicating a copy
+                /// from input shape at the same index.
+                Reshape(const Output<Node>& arg,
+                        const Output<Node>& pattern,
+                        bool zero_flag = false);
+
+                void validate_and_infer_types() override;
+
+                size_t get_version() const override { return 1; }
+                virtual std::shared_ptr<Node>
+                    copy_with_new_args(const NodeVector& new_args) const override;
+
+                bool get_zero_flag() const { return m_zero_flag; }
+                void set_zero_flag(bool zero_flag) { m_zero_flag = zero_flag; }
+            protected:
+                virtual void generate_adjoints(autodiff::Adjoints& adjoints,
+                                               const NodeVector& deltas) override;
+
+            private:
+                bool m_zero_flag;
+            };
+        }
     }
 }
